refactor(fgumi-bam-io): split bam_io into reader, writer, and paths modules
Splits the monolithic bam_io module into three focused modules:
- reader: BgzfReaderEnum, BamReaderAuto, RawBamReaderAuto, PipelineReaderOpts,
          and all create_bam_reader* / create_raw_bam_reader* factories
- writer: BgzfWriterEnum, BamWriter, RawBamWriter, IndexingBamWriter,
          and all create_*_writer / write_bai_index factories
- paths: is_stdin_path and is_stdout_path helpers

Re-exports all public items from the crate root for a flat API surface.
Fixes header.rs test imports (crate::bam_io -> crate::writer).
